requirements.txt
Scrapy==0.24.0
Twisted==15.4.0
argparse==1.2.1
backports.ssl-match-hostname==3.4.0.2
cffi==1.2.1
characteristic==14.3.0
cryptography==1.0.1
cssselect==0.9.1
decorator==4.0.2
enum34==1.0.4
idna==2.0
ipaddress==1.0.14
ipython==4.0.0
ipython-genutils==0.1.0
lxml==3.4.4
path.py==8.1.1
pexpect==3.3
pickleshare==0.5
pushbullet.py==0.8.1
pyOpenSSL==0.15.1
pyasn1==0.1.8
pyasn1-modules==0.0.7
pycparser==2.14
python-decouple==3.0
python-magic==0.4.6
queuelib==1.4.2
requests==2.7.0
service-identity==14.0.0
simplegeneric==0.8.1
six==1.9.0
traitlets==4.0.0
w3lib==1.12.0
websocket-client==0.32.0
wsgiref==0.1.2
zope.interface==4.1.2
